moved the ui components scripts and styles into separate directories. changed the readme file name
