+++ index.html
@@ -1,11 +1,39 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Landing Page</title>
+    <link rel="stylesheet" href="../styles/resets.css">
+    <link rel="stylesheet" href="../styles/main.css">
+    <link rel="stylesheet" href="../styles/component-styles/header.css">
+    <link rel="stylesheet" href="../styles/component-styles/navbar.css">
+    <link rel="stylesheet" href="../styles/component-styles/hero.css">
+    <link rel="stylesheet" href="../styles/component-styles/product-card.css">
+    <link rel="stylesheet" href="../styles/component-styles/landing.css">
 </head>
-<body>
-  
+
+<body class="landing">
+    <header id="header">
+    </header>
+    <navbar id="navbar">
+    </navbar>
+    <section id="hero">
+    </section>
+
+    <main id="main">
+        <section>
+            <h1>Latest Products</h1>
+        </section>
+        <section id="products-container">
+        </section>
+    </main>
+    <script type="module" src="../scripts/ui-components/header.js"></script>
+    <script type="module" src="../scripts/ui-components/navbar.js"></script>
+    <script type="module" src="../scripts/ui-components/hero.js"></script>
+    <script type="module" src="../scripts/ui-components/product-card.js"></script>
+
 </body>
+
 </html>
