--- misc/header.css
@@ -1,290 +1,0 @@
-/* Custom Properties (Variables) */
-:root {
-  /* Colors */
-  --purple-color: #fb2e86;
-  --serchBar: #FB2E86;
-  --heroSection: #F6F5FF;
-  --shop-slider-header: #fff;
-  --hero-section: #f8f5ff;
-  --hero-section-primary-color: #ff007f;
-  --hero-section-text-color: #222;
-  --text-main: #151875;
-  --text--prmary:#b0b0b5;
-  --text-label: #3F509E;
-  --text-input: #E7E6EF;
-  --text--side-bar: #7E81A2;
-  --product-title:#111C85;
-  --discount-price: #FF2AAA;
-}
-
-html,
-body {
-  width: 100%;
-  height: 100vh;
-  font-family: "Josefin Sans", sans-serif;
-}
-
-/* Utility Classes */
-.container {
-  max-width: 1280px;
-  margin: 0 auto;
-}
-
-/* Shared Component Styles */
-/* Header */
-/* Shop-slider page styling*/
-header {
-  /* background-color: var(--shop-slider-header); */
-  background-color: #fff;
-  padding: 1rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  /* width: 100%; */
-}
-/* nav-bar => site name*/
-.nav-bar{
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  gap: 20px;
-  align-items: center;
-  padding: 20px;
-}
-.nav-bar .site-name{
-  font-size: 25px;
-  text-decoration: none;
-  color: black;
-  width: auto;
-  margin-right: auto;
-  margin-left: 30px;
-  transition: 0.6s ease-in-out;
-}
-.nav-bar .site-name:hover{
-    transform: scale(1.2);
-    transform: 0.8s ease-in-out;
-    color: rgba(71, 73, 73, 0.716);
-  }
-/* nav-bar => site links*/
-.nav-bar-links {
-  width: 300px;
-  display: flex;
-  margin-left: 30px;
-  margin-right: 10px;
-  gap: 30px;
-  justify-content: center;
-}
-.nav-bar-links .home-page{
-  color: var(--purple-color);
-}
-.nav-bar-links a{
-  text-decoration: none;
-  color: black;
-  transition: 0.6s all ease-in-out;
-}
-.nav-bar-links .home-page:hover,
-.nav-bar-links a:hover{
-  transform: scale(1.2);
-  transform: .8s ease-in-out;
-  color: rgba(71, 73, 73, 0.716); 
-}
-.search-bar{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 200px;
-  height: 30px;
-  gap: 2px;
-}
-.search-bar input{
-  border: solid 3px rgba(40, 37, 37, 0.242);
-  border-radius: 5px;
-}
-
-.search-bar span{
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: solid 2px rgba(40, 37, 37, 0.242);
-  border-radius: 4px;
-  height: 20px;
-  width: 20px;
-  background-color: var(--purple-color);
-  color: white;
-  cursor: pointer;
-}
-/*Hero Section*/
-.hero-section{
-  background-color: var(--heroSection);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 2px solid transparent;
-  border-radius: 10px;
-  padding: 50px;
-  gap: 20px; 
-}
-.hero-section-img1{
-  width: 150px;
-  padding-bottom: 200px;
-}
-.hero-section-component{
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  text-align: left;
-  max-width: 500px;
-  padding: 20px;
-}
-.hero-section-component-p1{
-  font-size: 14px;
-  font-weight: bold;
-  padding-bottom: 10px;
-  color: var(--hero-section-primary-color);
-}
-.hero-section-component-h1{
-  font-size: 20px;
-  color: var(--hero-section-text-color);
-  padding-bottom: 10px;
-}
-.hero-section-component-p2{
-  font-size: 16px;
-  color:#666;
-}
-.hero-section-shop-btn{
-  background-color: var(--hero-section-primary-color);
-  color: white;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  margin-top: 10px;
-  transition: 0.6s all ease-in-out;
-}
-.hero-section-shop-btn:hover{
-  transform: scale(1.2);
-  transition: 0.8s all ease-in-out;
-}
-.hero-section-img2{
-  width: 400px;
-}
-/* body section style*/
-.body-section{
-  display: flex;
-  padding: 10px;
-  /* margin-right:600px; */
-}
-.body-section-div1{
-  margin-right: auto;
-  padding: 30px;
-
-}
-.body-section-div1 h1{
-  justify-content: flex-start;
-  font-size: 20px;
-  color: var(--text-main);
-}
-.body-section-div1 p{
-  color: var(--text--prmary);
-  line-height: 30px;
-}
-.body-section-div2{
-  padding: 20px;
-  padding-right: 40px;
-  color: var(--text-label);
-}
-.body-section-div2 input{
-  border: 2px solid var(--text-input);
-  border-radius: 10px;
-}
-.body-section-div2 select{
-  border: 2px solid var(--text-input);
-  border-radius: 10px;
-  font-size: 15px;
-}
-/* Products styling*/
-.section-products{
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  align-items: flex-start;
-  /* margin: 20px; */
-  /* margin-left: 50px; */
-  gap: 20px;
-}
-.side-bar-filter{
-  display: flex;
-  flex-direction: column;
-  width: 30%;
-  max-width: 300px;
-  padding-left: 100px;
-}
-.body-section-side-bar{
-  display: flex;
-  flex-direction: column;
-  /* align-items: flex-start; */
-  gap: 10px;
-  height: 100%;
-  padding-left: 40px;
-  margin-bottom: 30px;
-}
-.body-section-side-bar h2{
-  font-size: 20px;
-  text-decoration: underline;
-  color: var(--text-main);
-}
-.body-section-side-bar label{
-  color: var(--text--side-bar);
-  cursor: pointer;
-}
-.body-section-side-bar input{
-  background-color: var(--text--side-bar);
-}
-/*Products styling*/
-.products{
-  margin-left: 200px;
-  width: 70%;
-  display: flex;
-  flex-wrap: wrap;
-}
-.card{
-  display: flex;
-  background-color: var(--hero-section);
-  border-radius: 10px;
-  padding: 15px;
-  gap: 15px;
-  max-width: 1000px;
-}
-.card img{
-  width: 200px;
-  height: 200px;
-}
-.product-info{
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-.product-title{
-  font-size: 18px;
-  color: var(--product-title);
-  font-weight: bold;
-  padding-top: 10px;
-}
-.product-pricing{
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-.discounted-price{
-  color: var(--discount-price);
-  text-decoration: line-through;
-  font-weight: bold;
-}
-.original-price{
-  color: var(--product-title);
-  font-weight: bold;
-}
-.card p{
-  color: var(--text--side-bar);
-}
