--- pages/landing.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Landing Page</title>
-    <link rel="stylesheet" href="../styles/resets.css">
-    <link rel="stylesheet" href="../styles/main.css">
-    <link rel="stylesheet" href="../styles/header.css">
-    <link rel="stylesheet" href="../styles/navbar.css">
-    <link rel="stylesheet" href="../styles/hero.css">
-    <link rel="stylesheet" href="../styles/product-card.css">
-    <link rel="stylesheet" href="../styles/landing.css">
-</head>
-<body>
-    <div class="app">
-        <header id ="header">
-            <script type="module" src="../scripts/header.js"></script>
-        </header>
-        <navbar id = "navbar" >
-            <script type="module" src="../scripts/navbar.js"></script>
-        </navbar>
-        <section id = "hero">
-            <script type="module" src="../scripts/hero.js"></script>
-        </section>
-        
-        <main id = "main">
-            <section>
-                <h1>Latest Products</h1>
-            </section>
-            <section id="products-container">
-                <script type="module" src="../scripts/product-card.js" ></script>
-            </section>
-        </main>
-        
-    </div>
-</body>
-</html>
