+++ styles/component-styles/header.css
@@ -1,4 +1,4 @@
-@import url("./main.css");
+@import url("../main.css");
 
 .header {
     display: flex;
