+++ styles/component-styles/hero.css
@@ -1,4 +1,4 @@
-@import url("./main.css");
+@import url("../main.css");
 
 .hero {
     height: 764px;
