+++ styles/component-styles/landing.css
@@ -1,4 +1,4 @@
-@import url("./main.css");
+@import url("../main.css");
 
 
 h1 {
