+++ styles/component-styles/navbar.css
@@ -1,4 +1,4 @@
-@import url("./main.css");
+@import url("../main.css");
 
 .navbar {
     height: 78px;
